+++ src/Event.ts
@@ -1,5 +1,3 @@
-import type * as storage from "./StorageData";
-
 /**
  * イベントコード
  */
@@ -15,25 +13,6 @@ export const enum EventCode {
 	Operation = 0x40
 }
 
-/**
- * イベント共通のインターフェース。
- */
-export interface Event extends Array<any> {
-	[index: number]: any;
-	/**
-	 * @param EventCode
-	 */
-	0: EventCode;
-	/**
-	 * @param イベントフラグ
-	 */
-	1: number;
-	/**
-	 * @param プレイヤーID
-	 */
-	2: string | null;
-}
-
 export const enum EventIndex {
 	Code = 0,
 	EventFlags = 1,
@@ -46,21 +25,6 @@ export const enum EventFlagsMask {
 	Ignorable = 0b10000
 }
 
-/**
- * JoinEvent (0x0)
- * プレイヤーの参加イベント。
- */
-export interface JoinEvent extends Event {
-	/**
-	 * @param プレイヤー名
-	 */
-	3: string;
-	/*
-	 * @param ストレージデータ
-	 */
-	4?: storage.StorageData[];
-}
-
 export const enum JoinEventIndex {
 	Code = 0,
 	EventFlags = 1,
@@ -69,28 +33,11 @@ export const enum JoinEventIndex {
 	StorageData = 4
 }
 
-/**
- * LeaveEvent (0x1)
- * プレイヤーの離脱イベント。
- */
-export interface LeaveEvent extends Event {
-}
 
 export const enum LeaveEventIndex {
 	Code = 0,
 	EventFlags = 1,
 	PlayerId = 2
-}
-
-/**
- * TimestampEvent（0x2）
- * 時間の記録を表すイベント。
- */
-export interface TimestampEvent extends Event {
-	/**
-	 * @param 時間を表す値
-	 */
-	3: number;
 }
 
 export const enum TimestampEventIndex {
@@ -100,16 +47,6 @@ export const enum TimestampEventIndex {
 	Timestamp = 3
 }
 
-export interface PlayerInfoEvent extends Event {
-	/**
-	 * @param プレイヤー名
-	 */
-	3: string;
-	/**
-	 * @param ユーザ定義データ
-	 */
-	4?: any;
-}
 
 export const enum PlayerInfoEventIndex {
 	Code = 0,
@@ -119,16 +56,6 @@ export const enum PlayerInfoEventIndex {
 	UserData = 4
 }
 
-/**
- * MessageEvent (0x20)
- * 汎用的なデータを表すイベント。
- */
-export interface MessageEvent extends Event {
-	/**
-	 * @param データ
-	 */
-	3: any;
-}
 
 export const enum MessageEventIndex {
 	Code = 0,
@@ -137,32 +64,6 @@ export const enum MessageEventIndex {
 	Data = 3
 }
 
-/**
- * PointDownEvent (0x21)
- * ポイントダウンイベント。
- */
-export interface PointDownEvent extends Event {
-	/**
-	 * @param ポインターID
-	 */
-	3: number;
-	/**
-	 * @param X座標
-	 */
-	4: number;
-	/**
-	 * @param Y座標
-	 */
-	5: number;
-	/**
-	 * @param エンティティID
-	 */
-	6?: number;
-	/**
-	 * @param ボタンの種類
-	 */
-	7?: number;
-}
 
 export const enum PointDownEventIndex {
 	Code = 0,
@@ -175,48 +76,6 @@ export const enum PointDownEventIndex {
 	Button = 7
 }
 
-/**
- * PointMoveEvent (0x22)
- * ポイントムーブイベント。
- */
-export interface PointMoveEvent extends Event {
-	/**
-	 * @param ポインターID
-	 */
-	3: number;
-	/**
-	 * @param X座標
-	 */
-	4: number;
-	/**
-	 * @param Y座標
-	 */
-	5: number;
-	/**
-	 * @param ポイントダウンイベントからのX座標の差
-	 */
-	6: number;
-	/**
-	 * @param ポイントダウンイベントからのY座標の差
-	 */
-	7: number;
-	/**
-	 * @param 直前のポイントムーブイベントからのX座標の差
-	 */
-	8: number;
-	/**
-	 * @param 直前のポイントムーブイベントからのY座標の差
-	 */
-	9: number;
-	/**
-	 * @param エンティティID
-	 */
-	10?: number;
-	/**
-	 * @param ボタンの種類
-	 */
-	11?: number;
-}
 
 export const enum PointMoveEventIndex {
 	Code = 0,
@@ -231,49 +90,6 @@ export const enum PointMoveEventIndex {
 	PrevDeltaY = 9,
 	EntityId = 10,
 	Button = 11,
-}
-
-/**
- * PointUpEvent (0x23)
- * ポイントアップイベント。
- */
-export interface PointUpEvent extends Event {
-	/**
-	 * @param ポインターID
-	 */
-	3: number;
-	/**
-	 * @param X座標
-	 */
-	4: number;
-	/**
-	 * @param Y座標
-	 */
-	5: number;
-	/**
-	 * @param ポイントダウンイベントからのX座標の差
-	 */
-	6: number;
-	/**
-	 * @param ポイントダウンイベントからのY座標の差
-	 */
-	7: number;
-	/**
-	 * @param 直前のポイントムーブイベントからのX座標の差
-	 */
-	8: number;
-	/**
-	 * @param 直前のポイントムーブイベントからのY座標の差
-	 */
-	9: number;
-	/**
-	 * @param エンティティID
-	 */
-	10?: number;
-	/**
-	 * @param ボタンの種類
-	 */
-	11?: number;
 }
 
 export const enum PointUpEventIndex {
@@ -291,20 +107,6 @@ export const enum PointUpEventIndex {
 	Button = 11
 }
 
-/**
- * OperationEvent (0x40)
- * 操作プラグインイベント。
- */
-export interface OperationEvent extends Event {
-	/**
-	 * @param 操作プラグインコード
-	 */
-	3: number;
-	/**
-	 * @param 操作プラグインデータ
-	 */
-	4: (number|string)[];
-}
 
 export const enum OperationEventIndex {
 	Code = 0,
