+++ src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from "./Tick";
 export * from "./Event";
-export * from "./StorageData";
