+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./Tick";
 export * from "./Event";
 export * from "./StorageData";
-export * from "./constants/index";
+export * from "./constants";
