+++ tsconfig.json
@@ -14,9 +14,7 @@
     "./src/Event.ts",
     "./src/StorageData.ts",
     "./src/index.ts",
-    "./src/constants/Tick.ts",
     "./src/constants/Event.ts",
-    "./src/constants/StorageData.ts",
     "./src/constants/index.ts",
   ]
 }
